pure java main
